requirements.txt
# Core dependencies for inoculation x self-awareness experiment
openai>=1.37.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.4
pydantic>=2.7.0

# Analysis (only needed for plotting)
numpy
matplotlib
seaborn
pandas
adjustText>=1.2.0

# Testing
pytest>=8.0.0

# vLLM for local inference
vllm
